requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.46
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
bcrypt==4.1.1
marshmallow==3.20.1
python-dateutil==2.8.2
pytest==7.4.3
pytest-flask==1.3.0
pytest-celery==0.0.0
Flask-SQLAlchemy==3.1.1